requirements.txt
django
Pillow
django-embed-video
django-rosetta
django-modeltranslation
psycopg2